refactor: pages/_intrnal 配下の definitions と services を撲滅
